// app/interfaces/Book.ts
export default interface Book {
    title: string;
    author: string;
    cover: string;
    coverMedium: string;
    coverSmall: string;
    bookId: string;
    isbn: string;
    description: string;
    publishedYear: string;
    averageRating: string;
    userRating: string;
    userShelves: string;
    userReadAt: string;
    userDateAdded: string;
    numPages: string;
    link: string;
}